refactoring shared code into components
