+++ client/src/helpers/routes.ts
@@ -14,5 +14,6 @@ export const REGISTER_ORG = `${REGISTER}/${ORGANIZATION}`
 export const ACCOUNT_CREATE_ORG = `${ACCOUNT}${ORGANIZATION}`
 export const ACCOUNT_CHANGE_PASSWORD = `${ACCOUNT}/change-password`
 export const ACCOUNT_JOIN_ORGANIZATION = `${ACCOUNT}/join-organization`
+export const ACCOUNT_SWITCH_ORGANIZATION = `${ACCOUNT}/switch-organization`
 export const ACCOUNT_NOTIFICATION_SETTINGS = `${ACCOUNT}/notification-settings`
 export const NOTIFICATIONS = "/notifications"
