+++ client/src/pages/organization/OrganizationAddEditStatuses.tsx
@@ -1,0 +1,8 @@
+import React from "react"
+import { AddEditStatusForm } from "../../components/forms/AddEditStatusForm"
+
+export const OrganizationAddEditStatuses = () => {
+	return (
+		<AddEditStatusForm/>	
+	)
+}
